The EFI implementation does not like overlapping partition open/close
If a partition is opened/closed by more then one handle at any given
time then the runtime services get corrupted and the kernel crashes
during boot.

Since we have the partition mounted already in the only place we
absolutely have to have parse_config working, alter it to take
the pre-mounted partition as an argument instead of the handle
we passed it before.