Split argument parsing into two phases
Arguments can be broadly divided into two classes: those that
operate on the input and output files, and those that specify
what the input and output files are.

We don't want to impose command line ordering on the caller as this
amounts to forcing the user to understand the implementation details
of the bootconf tool, so we split the parsing into two phases instead.