steamcl-install: Rework lsblk output parsing in find_esp()
The issue with the current code (the `read < <(lsblk)` part) is that it
expects that lsblk will consistently fill every column with something.
However if ever lsblk leaves some column empty, then we get it wrong,
and we end up saving wrong values in the variables.

It turns out that lsblk might leave some columns empty, especially if
we're running in a container, udev is not really functional, and lsblk
can't get neither PARTTYPE, neither FSTYPE. (it usually gets those from
udev).

So we need a more robust parsing of lsblk output, and we can do that by
using the `--pairs` option.

----

Verify the issue with this snippet, see what happens when columns are
empty.

    lsblk -nlp -o NAME,PKNAME,PARTTYPE,FSTYPE,MOUNTPOINT

    while read part dev ptuid fs mount;
    do
        esp_part="$part";
        esp_dev="$dev";
        esp_uuid="$ptuid";
        esp_fs="$fs";
        esp_mount="$mount";

        echo -n "esp_part=$esp_part "
        echo -n "esp_dev=$esp_dev "
        echo -n "esp_uuid=$esp_uuid "
        echo -n "esp_fs=$esp_fs "
        echo -n "esp_mount=$esp_mount "
        echo
    done < <(lsblk -nlp -o NAME,PKNAME,PARTTYPE,FSTYPE,MOUNTPOINT);

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>