steamcl-install: Rework the coding style a bit in find_esp()
This is the same logic as before (AFAIU), but I reworked the style for
clarity. After this commit, the two read blocks are almost identical,
and the logic is obvious to the reader (well at least to me).

This is to prepare the next commit, which brings a third read block.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>