steamcl-install: Handle first install
When we install for the first time, the code path we follow is roughly:

  esp_boot_ok is 0
  pkg_is_newer is 1
  pkg_scum_ok is 1

So we head straight to update_esp(), however at this moment this
function only supports updating an already installed bootloader, and
fails if no boot entry is found.

This commit adds support for first time install.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>