steamcl-install: Allow boot entry creation to fail on system without efi vars
When we create a disk image with debos, there's no support for EFI
variables (and for EFI in general, as debos uses fakemachine, which in
turn is backed by a QEmu VM without EFI support).

In such environment, we can't create boot entries.

This commit adds a function to check if EFI variables are supported.
In case they're not, and boot entry creation failed, we just issue a
message, and we return successfully.

This commit DOES NOT address repetitive error messages from efibootmgr.
If we want a bit of a nicer output, we could bail out of functions that
call efibootmgr at the beginning, and avoir these error messages.

We could also add a command-line argument to the script, to explicitly
say if we're on a system without efi variables support (similar to
grub-install(8) `--no-nvram`).

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>