steamcl-install: Add an option to install the chainloader at boot/bootx64.efi
This is needed when we prepare a disk image with debos, and we want this
disk image to be bootable in QEmu at least, and everywhere at most. QEmu
won't boot the image if it doesn't find an efi binary at
boot/bootx64.efi.

The name for the option, --force-extra-removable, is taken from similar
option in grub-install.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>