bootload: Fix the sort loop (for real)
The first attempt to fix this loop was made at ee74c508, but it actually
introduced a bug.

The termination condition of the loop is indeed `i < j - 1`, because we
compare entries at `i` and `i + 1`. So we need to be sure that there is
an entry at `i + 1`. In the current version, when there is one entry
only at `i`, we compare it against an empty entry at `i + 1`, so
`found[ i + 1 ].at` is zero, hence we swap the only entry with an empty
entry.

So this is now the right fix: we run the sort algorithm only if there's
more than one entry, in this case we know that `j - 1` is > 0 and we can
safely use it as the termination condition.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>