EFI_DEVICE_PATH is a struct: passable instances should be pointers
In particular the UEFI call wrappers want an EFI_DEVICE_PATH **
into which they can place their own address out of the internally
managed memory pool of the UEFI services: The member of found_cfg
must be an EFI_DEVICE_PATH * for this to work.

UEFI does not typically throw an error if you pass a wrongly
indirected value here but you end up with a struct full of garbage
if you do.

We never noticed before because this member was for debugging
and not actually used (although it is about to be).