steamcl-install: Simplify (again) the logic in update_esp()
It's more readable to use bootdir, and then explicitly state the name of
the files steamcl.efi and steamcl.version, rather than using the
infamous construct `{bootpath%/*}` all over the place.

Additionally, there was a subtle bug lurking in the line:

    cp -av $bin $ver $esp${bootpath%/*};

where we assumed that $bin was named `version.efi`. A reasonable
assumption, but still, we can do better and make no assumption at all,
and explicitly install the file under the name steamcl.efi.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>