steamcl-install: Add support for --restrict
Note that the restrict flag file will be installed only if the steamcl
is installed. In case the steamcl is already up to date, we don't even
enter update_esp(), so --restrict will have no effect.

Also there is no way to unrestrict, apart from removing the restricted
file from the ESP by hand.

This is not great, and we might want to improve that in the future.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>