steamcl-install: Make sure /esp is mounted
This is something that is already done in the systemd service file that
starts this command, but it's not robust at all, it needs to be done
here and now if we want things to work.

For reference, and while testing a USB boot of SteamOS, the first
attempt to find the /esp fails (because lsblk doesn't find anything
mounted at /esp), and the second attempt, which only looks at the
partition type and the filesystem, find the ESP of the hard drive, not
the ESP of the USB drive.

It fails a bit further down the road, but heck, it's definitely not what
we want.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>