Improve mechanism to restrict bootloader selection
The mecanism to restrict the bootloader selection is not restrictive
enough.

The current implementation checks if the bootloader is located under the
same root element that the running chainloader image (i.e. it checks
only the first element of the path).

This commit iterates over every single path element until a hard-drive
is found. If a hard-drive is found, it considers the bootloader to be
restricted candidate. If not, the bootloader is not even considered as a
candidate.