Prompt text menu after three boot attempts
The second loader is responsible of flagging itself as invalid image and
reseting the chainloader boot attempts counter. If it does not and after
a certain amount of attemps, the chainloader consider that the automatic
path is possibly broken. In this situation, the chainloader could prompt
the user to select which the entry to boot.

This implements a text menu to prompt the selections of the discovered
entries to the user. additionally, this honors the two variables:
 - LoaderConfigTimeout that contains the boot menu timeout in
   microseconds.
 - LoaderTimeMenuUSec tells the userspace the time spent in the menu;
   systemd-analyse used it to analyze the boot time.

Note: A timeout of 0 microseconds turns off the timeout, and the menu is
shown indefinitely.

On Qemu with ed2k OVMF, bootctl reports the new informations:

	(root@steamos steamos)# bootctl --esp-path /esp
	(...)
	Current Boot Loader:
	      Product: steamcl 0.01
	     Features: ✗ Boot counting
	               ✓ Menu timeout control
	               ✗ One-shot menu timeout control
	               ✗ Default entry control
	               ✗ One-shot entry control
	               ✗ Support for XBOOTLDR partition
	               ✗ Support for passing random seed to OS
	               ✓ Boot loader sets ESP partition information