Create ChainLoader Entry Flags EFI variable
It is not possible to gather all informations about the path that leads
to the up and running system.

For example, the flag boot-other is lost on bootup. Additionnaly, the
chainloader forwards the flag update to the second loader using the EFI
options, even if flag is not set due to the logic of update-other.

The userspace may be interested in having this flags to report issues
about the boot.

This introduces the chainloader variable ChainloaderEntryFlags with the
new GUID 399abb9b-4bee-4a18-ab5b-45c6e0e8c716 to set flags both flags
boot-other and update.

The flags can be found using sysfs:

	(root@steamos steamos)# hexdump -C /sys/firmware/efi/efivars/ChainloaderEntryFlags-399abb9b-4bee-4a18-ab5b-45c6e0e8c716
	00000000  06 00 00 00 02 00 00 00  00 00 00 00              |............|
	0000000c