Create ChainLoader Info EFI variables
It is not possible to gather all informations about the path that leads
to the up and running system.

The path of the loader and its device UUID are stored in the variables
LoaderDevicePartUUID and LoaderImageIdentifier, but the second loader
can rewrite them if it implements the Boot Loader Interface as well
(i.e. if the second loader is systemd-boot or if GRUB implements the
Boot Loader Interface in the future).

This introduces the two chainloader variables ChainLoaderDevicePartUUID
and ChainLoaderImageIdentifier which are a duplication of the two loader
variables LoaderDevicePartUUID and LoaderImageIdentifier to make sure
the path to the chainloader is known by the userspace.

The infos can be found using sysfs:

	(root@steamos steamos)# hexdump -C /sys/firmware/efi/efivars/ChainLoaderDevicePartUUID-4a67b082-0a4c-41cf-b6c7-440b29bb8c4f
	00000000  06 00 00 00 34 00 34 00  37 00 39 00 32 00 34 00  |....4.4.7.9.2.4.|
	00000010  46 00 46 00 2d 00 44 00  37 00 33 00 36 00 2d 00  |F.F.-.D.7.3.6.-.|
	00000020  38 00 41 00 34 00 33 00  2d 00 41 00 32 00 32 00  |8.A.4.3.-.A.2.2.|
	00000030  32 00 2d 00 31 00 42 00  46 00 41 00 32 00 36 00  |2.-.1.B.F.A.2.6.|
	00000040  44 00 42 00 41 00 37 00  41 00 43 00 00 00        |D.B.A.7.A.C...|
	0000004e

	(root@steamos steamos)# hexdump -C /sys/firmware/efi/efivars/ChainLoaderImageIdentifier-4a67b082-0a4c-41cf-b6c7-440b29bb8c4f
	00000000  06 00 00 00 5c 00 45 00  46 00 49 00 5c 00 73 00  |....\.E.F.I.\.s.|
	00000010  74 00 65 00 61 00 6d 00  6f 00 73 00 5c 00 73 00  |t.e.a.m.o.s.\.s.|
	00000020  74 00 65 00 61 00 6d 00  63 00 6c 00 2e 00 65 00  |t.e.a.m.c.l...e.|
	00000030  66 00 69 00 00 00                                 |f.i...|
	00000036