Create ChainLoader Boot Attempts Counter
The second loader is responsible of flagging itself invalid. It it does
not because it is broken, it results in booting in loop forever.

This introduces the chainloader variable ChainLoaderBootAttempts which
is a counter that *has* to be reset by the second loader. For now, the
logic does nothing, but it should do something if the counter reaches
too many attempts without being reset.