Add support for LoaderConfigTimeoutOneShot
This honors the variable LoaderConfigTimeoutOneShot that contains a boot
menu timeout for the next following boot; it is automatically cleared
after being used and overrides LoaderConfigTimeout. A timeout of 0
microseconds turns off the timeout, and the menu is shown indefinitely.

The oneshot timeout can be created using sysfs:

	# printf "\x07\x00\x00\x00" >tmpfile
	# iconv -t utf-16le <<<"15000000" | tr '\n' '\0' >>tmpfile
	# cp tmpfile /sys/firmware/efi/efivars/LoaderConfigTimeoutOneShot-4a67b082-0a4c-41cf-b6c7-440b29bb8c4f
	# rm tmpfile

On Qemu with ed2k OVMF, bootctl reports the new informations:

	(root@steamos steamos)# bootctl --esp-path /esp
	(...)
	Current Boot Loader:
	      Product: steamcl 0.01
	     Features: ✗ Boot counting
	               ✓ Menu timeout control
	               ✓ One-shot menu timeout control
	               ✗ Default entry control
	               ✗ One-shot entry control
	               ✗ Support for XBOOTLDR partition
	               ✗ Support for passing random seed to OS
	               ✓ Boot loader sets ESP partition information