Reimplement function LibFileSystemVolumeLabelInfo
Some UEFI firmware are buggy:
 - they do not return a NULL-terminated string
 - they are confused with the sizes of UTF-16 characters

The second case is problematic as it results in buffer overflow while
writing the string to the pre-allocated buffer.

The firmware returns the number of characters written instead of the
number of bytes which is the 2 times less (i.e. sizeof(CHAR16) is 2 *
sizeof(char)). The function reallocates a new buffer based on that
returned value and calls the function again with a buffer that is too
narrow to contains the volume label ane the firmware results in buffer
overlow.

This reimplements the function LibFileSystemVolumeLabelInfo to handle
these buggy firmwares.