Choose -no-pie or -nopie in LDFLAGS depending on compiler support
The compiler may or may not have PIE output turned on, explicitly
or implicitly (and the distro may patch this). The linker should
be given the same instruction regarding PIE output.

The anti-PIE linker flag has changed (again, can be distro patched).

So we need to:
  - detect whether the compiler is invoked PIE mode (implicit or explicit)
  - disable PIE output for the ELF intermediate for EFI binaries
  - determine ehich linker anti-PIE flag(s) are recognised
  - pass linker anti-PIE flag(s) that match the compiler's invocation

NOTE: if you get the PIE-ness wrong you can produce an EFI binary that
looks superficially right (to `file' et al) but which simply does not run.

This fixes the error below:

	ld: Error: unable to disambiguate: -no-pie (did you mean --no-pie ?)