Don't automatically enable the menu based on chainloader boot count
It's possible that nothing ever gets around to resetting this counter
which would mean the menu was permanently enabled.

It's also not tremendously helpful since the A/B images shouldn't
actually _have_ EFI boot entries. It's the chainloader's job to
have an EFI boot entry and select which A/B/x image to boot.

The menu can be brought up by pressing <f3> during boot anyway.