Reduce exposure of the codebase to the underlying EFI library
Remove references to LibImageHandle, which we do not need.
Relocate the external initialisation code to util.c: The
goal is that all external EFI support code should reside
there so that changes to the EFI support have less impact on
the chainloader (and to make it easier to port to tianocore
instead, should we need to do so).