Sync other configs when setting boot modes (if necessary)
When steamos-bootconf is called with set-mode reboot-other or
set-mode update we set the boot-other flag on the target image.

Although the chainloader copes with the _other_ image(s) _also_ having
boot-other set, we should clear that flag on the other image config(s)
for clarity.

In reboot-other mode we also want to make sure the update flag is not
set on the image we expect to come up (for that you would want
"update" mode).

This does not prevent any arbitrary set of flags being set by calling
steamos-bootconf: It makes the results of the set-mode actions more
predictable.