If the fallback loader is steamcl, steamcl-install should update it
The default boot location (which the deck firmware, at least the
iteration I have) boots from is steamcl.efi then we keep it in
sync with /esp/efi/steamos/steamcl.efi.

If it is _not_ the steamos loader (eg another OS is co-installed
and it wants to be in charge) then we don't replace the fallback
loader.

The default loader is at efi/boot/bootx64.efi (on x86_64).