Use a guaranteed safe EFI file stat buffer size
We were allowing MAXFSNAMLEN for the filename but this turns out
to be a cargo culted value which could theoretically be breached
per the spec.

Although we would never break this limit in practice (or even get
close to it) we can allocate enough space to store the maximum
allowable VFAT file name length to be safe.