Check the comparative ages of old and new configs when migrating
Instead of just checking whether the new config location already
exists we now check that it is newer than the old location.

This is to handle cases where the OOBE image has been installed
on a partition, as it will update the _old_ location (since it
has the old bootconf layout and tools), and we don't want its
boot configuration changes to be ignored.