Add a safe UEFI command line fetch function get_image_cmdline
The UEFI command line data available from the LOADED_IMAGE protocol
is not guaranteed to be well formed or NUL terminated. It is simply
a blob of data with a known size.

It may also be read-only.

This fetcher returns a newly allocated copy of the data which is
guaranteed to end in L'\0' so it can safely be inspected and
manipulated by the normal string and wide-string functions.