Improved menu readability and more data + options.
The selected menu option is now marked with > … < labels as
well as being a different colour. Each image/config is also
labelled with its last boot time as well as having further
menu entries to boot in verbose mode or boot to the stage 2
(ie grub) menu.

The menu looks like this:

       A     Last Boot: 2022-02-09 17:45
       A     Verbose Boot
       A     OS Boot Menu
       DEV   Last Boot: 2022-02-09 18:12
    >  DEV   Verbose Boot                 <
       DEV   OS Boot Menu