Helper to check the state of the HW config button state EFI variable
The Deck firmware is going to provide a volatile EFI var to track
the state of the (⋯) button on boot (ie before the chainloader starts)
so that we can decide whether or not to display the boot menu.

This is because although we can detect keypresses from within the
chainloader itself we have only a short window to do so and it doesn't
catch keypresses that started _before the chainloader started:

       pre-boot | firmware | chainloader | …
    A     |-->| :          :             :
    B       |---:---->|    :             :
    C       |---:----------:-->|         :
    D           :  |--->|  :             :
    E           :  |-------:---->|       :
    F           :          :  |----->|   :
    G           :          :  |----------:--->|

The chainloader can use UEFI API to reliably detect cases F & G.
The firmware state variable allows us to catch states B,C,D,E as well.

The variable namespace and name are:

    f0393d2c-78a4-4bb9-af08-2932ca0dc11e
    JupiterFunctionConfigVariable