mark text_menu_choose_ as static
It's only used in this compilation unit.

For some reason, this is the difference between "works" and "hangs forever",
where "hangs forever" is right up to the return is all good, and then it never
gets back to the caller, and I don't know why because I can't hook the EFI
thing to a debugger.